Revise according to comments
